Extracted stubbed response classes to separate files
