+++ core/src/main/kotlin/in/specmatic/stub/HttpStub.kt
@@ -615,35 +615,6 @@ internal suspend fun respondToKtorHttpResponse(call: ApplicationCall, httpRespon
     call.respond(textContent)
 }
 
-interface StubbedResponseResult {
-    fun log(logs: MutableList<StubEndpoint>, httpRequest: HttpRequest)
-
-    val response: HttpStubResponse
-}
-
-class FoundStubbedResponse(override val response: HttpStubResponse) : StubbedResponseResult {
-    override fun log(logs: MutableList<StubEndpoint>, httpRequest: HttpRequest) {
-        logs.add(
-            StubEndpoint(
-                response.scenario?.path,
-                httpRequest.method,
-                response.response.status,
-                response.feature?.sourceProvider,
-                response.feature?.sourceRepository,
-                response.feature?.sourceRepositoryBranch,
-                response.feature?.specification,
-                response.feature?.serviceType,
-            )
-        )
-    }
-}
-
-class NotStubbed(override val response: HttpStubResponse) : StubbedResponseResult {
-    override fun log(logs: MutableList<StubEndpoint>, httpRequest: HttpRequest) {
-
-    }
-}
-
 fun getHttpResponse(
     httpRequest: HttpRequest,
     features: List<Feature>,
