+++ core/src/main/kotlin/in/specmatic/stub/StubbedResponseResult.kt
@@ -1,0 +1,10 @@
+package `in`.specmatic.stub
+
+import `in`.specmatic.core.HttpRequest
+import `in`.specmatic.stub.report.StubEndpoint
+
+interface StubbedResponseResult {
+    fun log(logs: MutableList<StubEndpoint>, httpRequest: HttpRequest)
+
+    val response: HttpStubResponse
+}
